dom/animation/test/css-transitions/test_animation-player-starttime.html

<!doctype html>
<html>
  <head>
    <meta charset=utf-8>
    <title>Tests for the effect of setting a CSS transition's
           AnimationPlayer.startTime</title>
    <style>

.animated-div {
  margin-left: 100px;
  transition: margin-left 1000s linear 1000s;
}

    </style>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="../testcommon.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script type="text/javascript">

'use strict';

// TODO: add equivalent tests without an animation-delay, but first we need to
// change the timing of animationstart dispatch. (Right now the animationstart
// event will fire before the ready Promise is resolved if there is no
// animation-delay.)
// See https://bugzilla.mozilla.org/show_bug.cgi?id=1134163

// TODO: Once the computedTiming property is implemented, add checks to the
// checker helpers to ensure that computedTiming's properties are updated as
// expected.
// See https://bugzilla.mozilla.org/show_bug.cgi?id=1108055


const ANIM_DELAY_MS = 1000000; // 1000s
const ANIM_DUR_MS = 1000000; // 1000s
const ANIM_PROPERTY_VAL = 'anim ' + ANIM_DUR_MS + 'ms ' + ANIM_DELAY_MS + 'ms';

/**
 * These helpers get the value that the startTime needs to be set to, to put an
 * animation that uses the above ANIM_DELAY_MS and ANIM_DUR_MS values into the
 * middle of various phases or points through the active duration.
 */
function startTimeForBeforePhase(timeline) {
  return timeline.currentTime - ANIM_DELAY_MS / 2;
}
function startTimeForActivePhase(timeline) {
  return timeline.currentTime - ANIM_DELAY_MS - ANIM_DUR_MS / 2;
}
function startTimeForAfterPhase(timeline) {
  return timeline.currentTime - ANIM_DELAY_MS - ANIM_DUR_MS - ANIM_DELAY_MS / 2;
}
function startTimeForStartOfActiveInterval(timeline) {
  return timeline.currentTime - ANIM_DELAY_MS;
}
function startTimeForFiftyPercentThroughActiveInterval(timeline) {
  return timeline.currentTime - ANIM_DELAY_MS - ANIM_DUR_MS * 0.5;
}
function startTimeForEndOfActiveInterval(timeline) {
  return timeline.currentTime - ANIM_DELAY_MS - ANIM_DUR_MS;
}


// Expected computed 'margin-left' values at points during the active interval:
// When we assert_between_inclusive using these values we could in theory cause
// intermittent failure due to very long delays between paints, but since the
// active duration is 1000s long, a delay would need to be around 100s to cause
// that. If that's happening then there are likely other issues that should be
// fixed, so a failure to make us look into that seems like a good thing.
const INITIAL_POSITION = 100;
const TEN_PCT_POSITION = 110;
const FIFTY_PCT_POSITION = 150;
const END_POSITION = 200;

/**
 * CSS animation events fire asynchronously after we set 'startTime'. This
 * helper class allows us to handle such events using Promises.
 *
 * To use this class:
 *
 *   var eventWatcher = new EventWatcher(watchedNode, eventTypes);
 *   eventWatcher.waitForEvent(eventType).then(function() {
 *     // Promise fulfilled
 *     checkStuff();
 *     makeSomeChanges();
 *     return eventWatcher.waitForEvent(nextEventType);
 *   }).then(function() {
 *     // Promise fulfilled
 *     checkMoreStuff();
 *     eventWatcher.stopWatching(); // all done - stop listening for events
 *   });
 *
 * This class will assert_unreached() if an event occurs when there is no
 * Promise created by a waitForEvent() call waiting to be fulfilled, or if the
 * event is of a different type to the type passed to waitForEvent. This helps
 * provide test coverage to ensure that only events that are expected occur, in
 * the correct order and with the correct timing. It also helps vastly simplify
 * the already complex code below by avoiding lots of gnarly error handling
 * code.
 */
function EventWatcher(watchedNode, eventTypes)
{
  if (typeof eventTypes == 'string') {
    eventTypes = [eventTypes];
  }

  var waitingFor = null;

  function eventHandler(evt) {
    if (!waitingFor) {
      assert_unreached('Not expecting event, but got: ' + evt.type +
        ' targeting element #' + evt.target.getAttribute('id'));
      return;
    }
    if (evt.type != waitingFor.types[0]) {
      assert_unreached('Expected ' + waitingFor.types[0] + ' event but got ' +
                        evt.type + ' event');
      return;
    }
    if (waitingFor.types.length > 1) {
      // Pop first event from array
      waitingFor.types.shift();
      return;
    }
    // We need to null out waitingFor before calling the resolve function since
    // the Promise's resolve handlers may call waitForEvent() which will need
    // to set waitingFor.
    var resolveFunc = waitingFor.resolve;
    waitingFor = null;
    resolveFunc(evt);
  }

  for (var i = 0; i < eventTypes.length; i++) {
    watchedNode.addEventListener(eventTypes[i], eventHandler);
  }

  this.waitForEvent = function(type) {
    if (typeof type != 'string') {
      return Promise.reject('Event type not a string');
    }
    return this.waitForEvents([type]);
  };

  /**
   * This is useful when two events are expected to fire one immediately after
   * the other. This happens when we skip over the entire active interval for
   * instance. In this case an 'animationstart' and an 'animationend' are fired
   * and due to the asynchronous nature of Promise callbacks this won't work:
   *
   *   eventWatcher.waitForEvent('animationstart').then(function() {
   *     return waitForEvent('animationend');
   *   }).then(...);
   *
   * It doesn't work because the 'animationend' listener is added too late,
   * because the resolve handler for the first Promise is called asynchronously
   * some time after the 'animationstart' event is called, rather than at the
   * time the event reaches the watched element.
   */
  this.waitForEvents = function(types) {
    if (waitingFor) {
      return Promise.reject('Already waiting for an event');
    }
    return new Promise(function(resolve, reject) {
      waitingFor = {
        types: types,
        resolve: resolve,
        reject: reject
      };
    });
  };

  this.stopWatching = function() {
    for (var i = 0; i < eventTypes.length; i++) {
      watchedNode.removeEventListener(eventTypes[i], eventHandler);
    }
  };

  return this;
}

// The terms used for the naming of the following helper functions refer to
// terms used in the Web Animations specification for specific phases of an
// animation. The terms can be found here:
//
//   http://w3c.github.io/web-animations/#animation-node-phases-and-states
//
// Note the distinction between "player start time" and "animation start time".
// The former is the start of the start delay. The latter is the start of the
// active interval. (If there is no delay, they are the same.)

// Called when the ready Promise's callbacks should happen
function checkStateOnReadyPromiseResolved(player)
{
  assert_less_than_equal(player.startTime, player.timeline.currentTime,
    'AnimationPlayer.startTime should be less than the timeline\'s ' +
    'currentTime on the first paint tick after animation creation');

  assert_equals(player.playState, 'running',
    'AnimationPlayer.playState should be "running" on the first paint ' +
    'tick after animation creation');

  var div = player.source.target;
  var marginLeft = parseFloat(getComputedStyle(div).marginLeft);
  assert_equals(marginLeft, INITIAL_POSITION,
                'the computed value of margin-left should be unaffected ' +
                'by an animation with a delay on ready Promise resolve');
}

// Called when startTime is set to the time the active interval starts.
function checkStateAtActiveIntervalStartTime(player)
{
  // We don't test player.startTime since our caller just set it.

  assert_equals(player.playState, 'running',
    'AnimationPlayer.playState should be "running" at the start of ' +
    'the active interval');

  var div = player.source.target;
  var marginLeft = parseFloat(getComputedStyle(div).marginLeft);
  assert_between_inclusive(marginLeft, INITIAL_POSITION, TEN_PCT_POSITION,
    'the computed value of margin-left should be close to the value at the ' +
    'beginning of the animation');
}

function checkStateAtFiftyPctOfActiveInterval(player)
{
  // We don't test player.startTime since our caller just set it.

  var div = player.source.target;
  var marginLeft = parseFloat(getComputedStyle(div).marginLeft);
  assert_equals(marginLeft, FIFTY_PCT_POSITION,
    'the computed value of margin-left should be half way through the ' +
    'animation at the midpoint of the active interval');
}

// Called when startTime is set to the time the active interval ends.
function checkStateAtActiveIntervalEndTime(player)
{
  // We don't test player.startTime since our caller just set it.

  assert_equals(player.playState, 'finished',
    'AnimationPlayer.playState should be "finished" at the end of ' +
    'the active interval');

  var div = player.source.target;
  var marginLeft = parseFloat(getComputedStyle(div).marginLeft);
  assert_equals(marginLeft, END_POSITION,
    'the computed value of margin-left should be the final transitioned-to ' +
    'value at the end of the active duration');
}

test(function(t)
{
  var div = addDiv(t, {'class': 'animated-div'});
  flushComputedStyle(div);
  div.style.marginLeft = '200px'; // initiate transition

  var player = div.getAnimationPlayers()[0];
  assert_equals(player.startTime, null, 'startTime is unresolved');
}, 'startTime of a newly created transition is unresolved');


test(function(t)
{
  var div = addDiv(t, {'class': 'animated-div'});
  flushComputedStyle(div);
  div.style.marginLeft = '200px'; // initiate transition

  var player = div.getAnimationPlayers()[0];
  var currentTime = player.timeline.currentTime;
  player.startTime = currentTime;
  assert_approx_equals(player.startTime, currentTime, 0.0001, // rounding error
    'Check setting of startTime actually works');
}, 'Sanity test to check round-tripping assigning to new animation\'s ' +
   'startTime');


async_test(function(t) {
  var div = addDiv(t, {'class': 'animated-div'});
  var eventWatcher = new EventWatcher(div, 'transitionend');

  flushComputedStyle(div);
  div.style.marginLeft = '200px'; // initiate transition

  var player = div.getAnimationPlayers()[0];

  player.ready.then(t.step_func(function() {
    checkStateOnReadyPromiseResolved(player);

    player.startTime = startTimeForStartOfActiveInterval(player.timeline);
    checkStateAtActiveIntervalStartTime(player);

    player.startTime =
      startTimeForFiftyPercentThroughActiveInterval(player.timeline);
    checkStateAtFiftyPctOfActiveInterval(player);

    player.startTime = startTimeForEndOfActiveInterval(player.timeline);
    return eventWatcher.waitForEvent('transitionend');
  })).then(t.step_func(function() {
    checkStateAtActiveIntervalEndTime(player);

    eventWatcher.stopWatching();
  })).catch(t.step_func(function(reason) {
    assert_unreached(reason);
  })).then(function() {
    t.done();
  });
}, 'Skipping forward through animation');


test(function(t) {
  var div = addDiv(t, {'class': 'animated-div'});
  var eventWatcher = new EventWatcher(div, 'transitionend');

  flushComputedStyle(div);
  div.style.marginLeft = '200px'; // initiate transition

  var player = div.getAnimationPlayers()[0];

  // Unlike in the case of CSS animations, we cannot skip to the end and skip
  // backwards since when we reach the end the transition effect is removed and
  // changes to the AnimationPlayer object no longer affect the element. For
  // this reason we only skip forwards as far as the 90% through point.

  player.startTime = startTimeForFiftyPercentThroughActiveInterval(player.timeline);
  checkStateAtFiftyPctOfActiveInterval(player);

  player.startTime = startTimeForStartOfActiveInterval(player.timeline);

  // Despite going backwards from being in the active interval to being before
  // it, we now expect an 'animationend' event because the animation should go
  // from being active to inactive.
  //
  // Calling checkStateAtActiveIntervalStartTime will check computed style,
  // causing computed style to be updated and the 'transitionend' event to
  // be dispatched synchronously. We need to call waitForEvent first
  // otherwise eventWatcher will assert that the event was unexpected.
  eventWatcher.waitForEvent('transitionend').then(function() {
    eventWatcher.stopWatching();
    t.done();
  });
  checkStateAtActiveIntervalStartTime(player);
}, 'Skipping backwards through transition');


async_test(function(t) {
  var div = addDiv(t, {'class': 'animated-div'});

  flushComputedStyle(div);
  div.style.marginLeft = '200px'; // initiate transition

  var player = div.getAnimationPlayers()[0];

  var storedCurrentTime;

  player.ready.then(t.step_func(function() {
    storedCurrentTime = player.currentTime;
    player.startTime = null;
    return player.ready;
  })).catch(t.step_func(function(reason) {
    assert_unreached(reason);
  })).then(function() {
    assert_equals(player.currentTime, storedCurrentTime,
      'Test that hold time is correct');
    t.done();
  });
}, 'Setting startTime to null');


async_test(function(t) {
  var div = addDiv(t, {'class': 'animated-div'});

  flushComputedStyle(div);
  div.style.marginLeft = '200px'; // initiate transition

  var player = div.getAnimationPlayers()[0];

  player.ready.then(t.step_func(function() {
    var savedStartTime = player.startTime;

    assert_not_equals(player.startTime, null,
      'AnimationPlayer.startTime not null on ready Promise resolve');

    player.pause();
    return player.ready;
  })).then(t.step_func(function() {
    assert_equals(player.startTime, null,
      'AnimationPlayer.startTime is null after paused');
    assert_equals(player.playState, 'paused',
      'AnimationPlayer.playState is "paused" after pause() call');
  })).catch(t.step_func(function(reason) {
    assert_unreached(reason);
  })).then(function() {
    t.done();
  });
}, 'AnimationPlayer.startTime after paused');

    </script>
  </body>
</html>
